Fix subtle race condition
When a subscription is made it was possible to miss out on a
notification:

Subscribe triggered API request and afterwards registered for changes.
A change coming in inbetween was not noticed.

This is fixed by not requesting the API in addRequest, but instead just
add the request in a way that a change notification will be triggered.